Pulled out analyzers to opensearch module
