+++ core/src/main/java/org/opensearch/sql/storage/StorageEngine.java
@@ -8,6 +8,10 @@ package org.opensearch.sql.storage;
 import java.util.Collection;
 import java.util.Collections;
 import org.opensearch.sql.DataSourceSchemaName;
+import org.opensearch.sql.analysis.Analyzer;
+import org.opensearch.sql.analysis.ExpressionAnalyzer;
+import org.opensearch.sql.datasource.DataSourceService;
+import org.opensearch.sql.expression.function.BuiltinFunctionRepository;
 import org.opensearch.sql.expression.function.FunctionResolver;
 
 /** Storage engine for different storage to provide data access API implementation. */
@@ -23,5 +27,9 @@ public interface StorageEngine {
    */
   default Collection<FunctionResolver> getFunctions() {
     return Collections.emptyList();
+  }
+
+  default Analyzer getAnalyzer(DataSourceService dataSourceService, BuiltinFunctionRepository repository) {
+    return new Analyzer(new ExpressionAnalyzer(repository), dataSourceService, repository);
   }
 }
